Refactors SymbolicTransformation to add a context that permits access to the Model->ModelType mapping via a new interface SymbolicTransformationContext.
To allow transformers to make use of this it has also been necessary to add
explicit access to the model through the Constraint interface.  Every
Constraint, and any resolver that was using getElement(3) has been updated to 
make a call to getModel() instead.

The explicit localization code in LocalQueryResolver has been factored into
ConstraintOperations.  This allows the LocalQueryResolver to correctly localize
and dispatch custom constraints to their appropriate resolver.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@93 1b8e2379-6217-0410-9814-82e2034ecad4