http://mulgara.org/jira/browse/MGR-2 "Multiple writers not supported" transactioning error occurs if reading and writing concurrently
This revision should fix MGR-2.

I believe the bug was the result of a race condition in StringPoolSession.  It
is possible for multiple transactions to enter prepare at the same time -
specifically a read-only transaction can enter prepare while a write transaction
is still flushing data to disk.  StringPoolSession was designed to detect
concurrent prepares and abort the latter attempts to enter the prepare state.

It is not sufficient to simply stop attempting to detect concurrent prepares,
as the store layer only permits a single call to prepare prior to a commit.
Removing the test leads to multiple possible prepares within a single
transaction, and so there is a need to add extra code to ensure this doesn't
happen.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@106 1b8e2379-6217-0410-9814-82e2034ecad4