The only methods left on DatabaseSession that are not directly involved in implementing the Session interface are
1) Transaction Methods - which will be replaced by the new MulgaraTransaction*
classes.

2) clearcache and fineResolverFactory - The former will be migrated to
OperationContextXAResource; the latter will be merged into
DatabaseOperationContext.

This commit also imports the MulgaraTransaction* classes and support.  This code
has been fixed from the xafix-design to the extent required to allow it to
compile.  It has also had various lines commented out to isolate it from the
existing code - permitting the existing code to continue to run unaffected.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@113 1b8e2379-6217-0410-9814-82e2034ecad4