This fixes MGR-44 - Error during transaction.commit() causes infinite recursion and looping
The transaction code is still beta, and indeed there are reference counting issues that need resolving in the presence of transaction failures.  There are still warnings being printed by the finalize() method reflecting a failure to dereference adequately from TransactionalAnswer.sessionClose().  This however is a seperate issue to MGR-44, and this fix (moving from a DFA based transaction model to a DPA transaction model) resolves the infinite recursion and makes a fix of the dereferencing issues in sessionClose possible.


git-svn-id: http://mulgara.org/svn/mulgara/trunk@207 1b8e2379-6217-0410-9814-82e2034ecad4