Adds a test fo AdvDatabaseSessionUnitTest that checks that delete/select implicitly commits. This is a repeatable test for MGR-48 (http://mulgara.org/jira/browse/MGR-48) "update results incorrect when autocommit on and multiple sessions are used".
We retain the patch provided by Ronald, as this bug was triggered by the delete operation
retaining a reference to the write-transaction until the garbarge-collector collected the
result of the inner select.  The result was a failure to commit, and therefore incorrect
results with concurrent sessions.  By closing the temporary answer, the reference count
was corrected and the transaction properly committed and terminated.

To ensure that this cannot happen again, several alternatives were considered that would have
provided an explicit call to 'commit' at the end of the delete.  These were rejected as this
would have required DatabaseSession to either track or query the auto-commit status to determine
if the transaction should be committed.  This would also bypass the reference counting 
maintained by the transaction object itself, and reintroduce transaction-handling logic into
DatabaseSession after this had been delegated to the 3 transaction management classes.

The ultimate solution chosen was to track the autocommit status explicitly within the manager, 
and to check this status when a transaction attempts to suspend.  The error we need to detect
is a write-transaction outliving an operation that is supposed to be atomic.  The only way a
transaction can outlive _any_ operation is to be suspended, so checking for validity on suspend
will guarantee this type of error can never occur.  The result will now be a failed atomic write
operation, something that will be highly visible and easily fixed.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@208 1b8e2379-6217-0410-9814-82e2034ecad4