Deleted ObjectPool and all references to it.
This is a real architectural change, as the use of these pools was pervasive in the lower layers.

In JDK 1.4, object pools were required for efficient operation, however, improvements in memory
management in JDK 1.5 changed the recommendations from Sun to eliminate object pools partly
to help the garbage collector, and partly because allocation was typically much faster:
  http://java.sun.com/docs/hotspot/HotSpotFAQ.html#gc_pooling
Since we no longer support JDK 1.4, then this is an overdue change. Apparently JDK 1.6 offers
even better performance.

A number of changes resulted in poor formatting, so this was cleaned up at the same time.
Also, a number of warnings were removed (e.g. non-typed WeakReferences and phase lists
in the FreeList class). This was outside of the scope of removing the object pools, but
the opportunity to make the modification was there, and these classes are unlikely to be
revisited in the foreseeable future.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@1071 1b8e2379-6217-0410-9814-82e2034ecad4