Restructured this file to treat each of the components in a more transparent way.
Now the main() method simply loads the configuration, and uses this to
instantiate an EmbeddedMulgaraServer instance. The server creates a ServerMBean
for building the Database along with an RMI service. It then checks if an http
server is required, and if so it creates a Jetty server, and registers the
known web applications. 

The main method then starts the EmbeddedMulgaraServer, which in turn starts the
Mulgara and HTTP servers.

The EmbeddedMulgaraServer object also used to double as a Thread so that it
could be registered as a shutdown handler. This functionality has been moved
into a new inner class devoted to this function.

Since the EmbeddedMulgaraServer is never accessed from outside of its own
main() method, many of the public methods have been changed to private. This
helped define what was being called and what wasn't, allowing many unnecessary
methods to be eliminated. Also, most of the public accessors were used for
accessing static data that was duplicated in ServerInfo. This duplication has
been removed, with ServerInfo now being the sole location for this data. The
accessors for this data were also removed.

Ultimately, this class should be replaced with one that has flexible
configuration for each possible service, with each service being attached to
shared Database instances (instead of an RMIServer instance). When multiple
services are made possible in this way, ServerInfo will no longer be static
and will be attached to a Database, instead of holding global data as it does
now.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@1115 1b8e2379-6217-0410-9814-82e2034ecad4