This is an integrated checkin, since each modification needs the others in order to work.
EmbeddedMulgaraServer has been extensively reshaped. This is not its final
form, but hopefully it is easier to understand and refactor.

All static documents have been removed from the web server. This may result
in the tutorial in the webui needing to be updated since the documents to be
loaded may be elsewhere. The web server can also be disabled, either by
setting <Disabled>false</Disabled> in the Jetty section of the config file
or by defining "mulgara.http.disable" in the system properties.

The web server has also been updated from Jetty 4.2.19 to Jetty 6.1.11. This
resulted in changes to the Jetty configuration. For details see:
  * conf/mulgara-embedded.dtd (not actually used anywhere)
  * conf/mulgara-embedded.xsd (the definition)
  * conf/mulgara-config.xml (The default configuration)

The library updates also required a few updates to the build scripts in
"descriptor" and "tag".

Also updated the main build scripts to NOT delete the SableCC and
ideSupport.jar files when building the "clean" target. If you want to remove
these, then use the new "mrclean" target.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@1118 1b8e2379-6217-0410-9814-82e2034ecad4