Fix concurrent string-pool-session access bug.
The parser, running in a separate thread, was allocating blank-node ids via
ResolverSession.localize(), thereby creating a situation where two threads
were accessing the same ResolverSession (and therefore the same
StringPoolSession) concurrently. The fix here is to delay the id allocation
and have the main thread do it instead.

Other solutions would have included adding extra synchronization to various
classes (e.g. the ResolverSession and StringPoolSession), but it seemed like
we'd be punishing all other code and paths for this one errant usage.


git-svn-id: http://mulgara.org/svn/mulgara/trunk@1143 1b8e2379-6217-0410-9814-82e2034ecad4