Fix backwards compatibility support added in r1144.
Pre r1081 the date was being stored in local-time, not UTC, i.e. as
midnight + local-time-offset. Reading that as UTC was then causing
delete-select to fail because globalization would return a pure date
with tz-offset 0, but subsequent localization would then not find the
db entry since the entry has a non-zero time (unless your locale happened
to be UTC). So now we figure out the tz-offset of the stored value and
use that instead.


git-svn-id: http://mulgara.org/svn/mulgara/trunk@1176 1b8e2379-6217-0410-9814-82e2034ecad4