Imported JRDF into the source tree, removing it from the library. Added generics to remove warnings.
This import has been a common request. It will provide easy access to view the source. It will also
allow for easy modification for fixing bugs, or extension where we have a need to expand the
definition. An example of this was the need to add generics, but we have also wanted to work with
JRDF objects and other objects (like Variables) using some kind of common interface.

The import as allowable as JRDF is licensed mostly with Apache 1.1. However, the RDF/XML parsing
code was imported from Sesame and was under LGPL. This code had to be removed which in turn had
an impact on the RIO content handler. Since this content handler was only being used in the unit
tests for the Relational Resolver, the content handler was removed.

While debugging the new integration some utilities were created and updated. This included a
Classpath utility, for printing the classpath and copying out jars to a temporary file, and 
moving the "functional" classes to their own package.



git-svn-id: http://mulgara.org/svn/mulgara/trunk@1199 1b8e2379-6217-0410-9814-82e2034ecad4