Adding an option to LiteralTuples to throw an exception if the tuples is closed more than once, which will match the behavior of HybridTuples.
This is needed to debug and test the TuplesOperations class.  However, there is also real code that uses this class, so I'm not comfortable changing the default behavior.  Therefore I'm adding a new constructor to expose this option; calls to the old constructor will have the existing behavior where multiple closes are allowed.

git-svn-id: http://mulgara.org/svn/mulgara/trunk@1300 1b8e2379-6217-0410-9814-82e2034ecad4