Make lucene-resolver transaction aware.
With this commit not only are changes committed or rolled back as appropriate,
but changes made in one transaction are not visible to other transactions till
they are committed, and then only to new transactions. This basically provides
snapshot isolation, but with one major caveat: changes are also not visible to
the transaction that made the change! E.g. a query following an insert in the
same transaction will not return any results that include the just inserted
data (in the lucene-resolver).

One possible way around the above caveat may be to create a new index for every
transaction, and then on a commit merge the index back into the main index; it
is unclear right now how costly that would be in terms of performance, though.

Because IndexReader's and IndexWriter's are now kept open for the duration of
the transaction, applications that do multiple queries (and/or writes) in a
transaction should see a noticeable performance improvement.

git-svn-id: http://mulgara.org/svn/mulgara/trunk@1311 1b8e2379-6217-0410-9814-82e2034ecad4