Added support for returning the score in fulltext (lucene) queries.
For this a new query graph is introduced with two new predicates, mulgara:search
and mulgara:score . As an example, the query

  select $foo from ... where $foo <title> 'sequencing' in <lucene-graph>;

now becomes

  select $foo $score from ... where
      $foo <mulgara:search> $search in <lucene-graph> and
                            $search <title> 'sequencing' in <lucene-graph> and
                            $search <mulgara:score> $score  in <lucene-graph>;

The score is optional; also, the previous simple form is still recognized. This
query graph is extensible, e.g. to add query parameters such as the analyzer to
use.

git-svn-id: http://mulgara.org/svn/mulgara/trunk@1366 1b8e2379-6217-0410-9814-82e2034ecad4